requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.10
django-cleanup==5.2.0
django-crispy-forms==1.13.0
django-email-log==1.0.0
django-json-widget==1.1.1
django-jsonfield==1.4.1
django-private-storage==3.0
future==0.18.2
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
isodate==0.6.1
oauth2client==4.1.3
oauthlib==3.1.1
Pillow==8.4.0
protobuf==3.19.1
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
reportlab==3.6.5
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
webencodings==0.5.1
xhtml2pdf==0.2.5
